Refactor dashboard workspace components
